% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{sona_get_secrets}
\alias{sona_get_secrets}
\title{Get Username and Password from .Rprofile}
\usage{
sona_get_secrets()
}
\description{
View the username and password you set up via sona_set_secrets(). Mainly an internal
function.
}
