% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SONA_functions.R
\name{sona_getstudy}
\alias{sona_getstudy}
\title{Access The Study List}
\usage{
sona_getstudy(main_url = "https://usu.sona-systems.com", active = -1,
  approved = -1, web_flag = -1, survey_flag = -1)
}
\arguments{
\item{main_url}{main URL of the sona system you are using}

\item{active}{should all studies be included (-1), just the active ones (1), or just the inactive ones (0)}

\item{approved}{should all studies be included (-1), just the approved ones (1), or just the un-approved ones (0)}

\item{web_flag}{should all studies be included (-1), just the web ones (1), or just the lab ones (0)}

\item{survey_flag}{should all studies be included (-1), just the online surveys ones (1), or all non-online survey ones (0)}
}
\description{
Retrieves the list of studies in the system based on study type.
}
