% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{sona_set_secrets}
\alias{sona_set_secrets}
\title{Securely Set Username and Password}
\usage{
sona_set_secrets(username, password)
}
\arguments{
\item{username}{your administrator username for SONA-Systems}

\item{password}{your administrator password for SONA-Systems tied to the username}
}
\description{
Securely tell R what your username and password is by calling this function once.
It saves the information in a hidden file (.Rprofile). sona_get_secrets() then
looks in this file for your information so that you can share code without
accidentally sharing private information.
}
