% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{sona_unset_secrets}
\alias{sona_unset_secrets}
\title{Delete Username and Password from .Rprofile}
\usage{
sona_unset_secrets()
}
\description{
Remove the information from .Rprofile that were set in sona_set_secrets().
}
